change the navigation structure and added SettingsScreen to nav bar
