+++ App.js
@@ -1,12 +1,5 @@
 import React from "react";
-import { NavigationContainer } from "@react-navigation/native";
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import LoginScreen from "./screens/LoginScreen";
-import SignupScreen from "./screens/SignupScreen";
-import StartScreen from "./screens/StartScreen";
-import QrCodeScannerScreen from "./screens/QrScannerScreen";
-import CommunityScreen from "./screens/CommunityScreen";
-import NavBar from "./components/NavBar";
+import Navigator from "./components/Navigator";
 import { LogBox } from "react-native";
 
 // Fonts
@@ -14,8 +7,6 @@ import { useFonts } from "expo-font";
 
 // suppressing a warning that is showing because of some default options in expo
 LogBox.ignoreLogs(["AsyncStorage"]);
-
-const Stack = createNativeStackNavigator();
 
 export default App = () => {
   // Loading fonts
@@ -28,21 +19,6 @@ export default App = () => {
   if (!fontsLoaded) {
     return null;
   } else {
-    return (
-      <NavigationContainer>
-        <Stack.Navigator
-          screenOptions={{
-            headerShown: false,
-          }}
-        >
-          <Stack.Screen name="StartScreen" component={StartScreen} />
-          <Stack.Screen name="App" component={NavBar} />
-          <Stack.Screen name="Login" component={LoginScreen} />
-          <Stack.Screen name="SignUp" component={SignupScreen} />
-          <Stack.Screen name="QrScanner" component={QrCodeScannerScreen} />
-          <Stack.Screen name="Community" component={CommunityScreen} />
-        </Stack.Navigator>
-      </NavigationContainer>
-    );
+    return <Navigator />;
   }
 };
