+++ package.json
@@ -13,16 +13,19 @@
     "@expo-google-fonts/bungee": "^0.2.0",
     "@expo-google-fonts/inter": "^0.2.0",
     "@fortawesome/react-native-fontawesome": "^0.2.7",
+    "@react-native-async-storage/async-storage": "~1.15.0",
     "@react-navigation/bottom-tabs": "^6.0.9",
     "@react-navigation/core": "^6.1.0",
+    "@react-navigation/drawer": "^6.1.8",
     "@react-navigation/native": "^6.0.6",
     "@react-navigation/native-stack": "^6.2.5",
     "@react-navigation/stack": "^6.0.11",
     "@use-expo/font": "^2.0.0",
     "expo": "~43.0.0",
     "expo-app-loading": "~1.2.1",
-    "expo-font": "~10.0.3",
     "expo-barcode-scanner": "~11.1.2",
+    "expo-font": "~10.0.3",
+    "expo-location": "~13.0.4",
     "expo-splash-screen": "~0.13.5",
     "expo-status-bar": "~1.1.0",
     "expo-updates": "~0.10.13",
@@ -38,9 +41,7 @@
     "react-native-screens": "~3.8.0",
     "react-native-svg": "12.1.1",
     "react-native-web": "0.17.1",
-    "react-native-webview": "^11.13.0",
-    "@react-native-async-storage/async-storage": "~1.15.0",
-    "expo-location": "~13.0.4"
+    "react-native-webview": "^11.13.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
