+++ screens/AboutUsScreen.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { View, Text } from "react-native";
+
+const AboutUsScreen = () => {
+  return (
+    <View>
+      <Text>About us Screen</Text>
+    </View>
+  );
+};
+
+export default AboutUsScreen;
