+++ screens/SettingsScreen.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { SafeAreaView, Text } from "react-native";
+
+const SettingsScreen = () => {
+  return (
+    <SafeAreaView>
+      <Text>Settings Screen</Text>
+    </SafeAreaView>
+  );
+};
+
+export default SettingsScreen;
